refactored code
refactored chart code
