+++ multi-manager-dashboard/src/components/charts/types.tsx
@@ -1,0 +1,13 @@
+export type CurveFitGraph = {
+    name: string;
+    data: CurveFitData[];
+};
+
+export type CurveFitData = {
+    index: number;
+    timestamp: number;
+    gain: number;
+    loss: number;
+    allocated: number;
+    duration: number;
+};
